requirements.txt
streamlit
yfinance
plotly
pandas
numpy
scipy
newsapi-python
streamlit-autorefresh